Create password generator
This program is a Python script that generates a random password of a given length. The program uses the random module to select random characters from a list of possible characters, such as letters, digits, and symbols. The program asks the user to enter the desired length of the password and then creates an empty string to store the password. The program then loops through the length of the password and appends a random character to the password string. Finally, the program prints the generated password to the screen. The program can be used to create strong and secure passwords for various purposes.